move shop appointments to shop redux
