+++ src/components/appointments/customer/customerAppointments.tsx
@@ -1,5 +1,5 @@
 import { AppointmentStatus, UserType } from "@prisma/client";
-import { readCustomerAppointments } from "@redux/actions/appointmentAction";
+import { readAppointments } from "@redux/actions/appointmentAction";
 import { AppointmentSelectors } from "@redux/selectors/appointmentSelectors";
 import { AuthSelectors } from "@redux/selectors/authSelectors";
 import { NextPage } from "next";
@@ -9,9 +9,7 @@ import { useDispatch, useSelector } from "react-redux";
 import { ICustomerAppointment } from "src/types/appointment";
 
 const CustomerAppointments: NextPage = () => {
-  const appointments = useSelector(
-    AppointmentSelectors.getCustomerAppointments
-  );
+  const appointments = useSelector(AppointmentSelectors.getAppointments);
   const [inProgressAppointments, setInProgressAppointments] = useState<
     ICustomerAppointment[]
   >([]);
@@ -47,7 +45,7 @@ const CustomerAppointments: NextPage = () => {
 
   useEffect(() => {
     if (customerId) {
-      dispatch(readCustomerAppointments({ customerId }));
+      dispatch(readAppointments({ id: customerId }));
     }
   }, [dispatch, customerId]);
