+++ src/redux/actions/shopActions.ts
@@ -1,5 +1,6 @@
 import { IShopState } from "@redux/state/shop/shopState";
 import ShopTypes from "@redux/types/shopTypes";
+import { IAppointment } from "src/types/appointment";
 import { IEmployee } from "src/types/employee";
 import { IService } from "src/types/service";
 
@@ -27,6 +28,10 @@ export interface IShopActionSetShopState extends IShopActionBase {
   payload: IShopState;
 }
 
+export interface IShopActionSetAppointments extends IShopActionBase {
+  payload: { appointments: IAppointment[] };
+}
+
 export type IShopAction =
   | IShopActionReadShopEmployees
   | IShopActionSetShopState
@@ -50,10 +55,22 @@ export const readShopServices = (payload: void) => ({
   payload,
 });
 
+export const readShopAppointments = (payload: void) => ({
+  type: ShopTypes.READ_SHOP_APPOINTMENTS,
+  payload,
+});
+
 export const setShopServices = (
   payload: IShopActionSetShopServices["payload"]
 ) => ({
   type: ShopTypes.SET_SHOP_SERVICES,
+  payload,
+});
+
+export const setShopAppointments = (
+  payload: IShopActionSetAppointments["payload"]
+) => ({
+  type: ShopTypes.SET_SHOP_APPOINTMENTS,
   payload,
 });
