+++ src/redux/sagas/shopSaga.ts
@@ -1,6 +1,9 @@
-import { Employee } from "@prisma/client";
+import { Appointment, Employee } from "@prisma/client";
 import { AuthSelectors } from "@redux/selectors/authSelectors";
 import ShopTypes from "@redux/types/shopTypes";
+import { getServiceById } from "@server/services/serviceService";
+import { getCustomerById } from "@server/services/userService";
+import { getVehicleById } from "@server/services/vehicleService";
 import {
   all,
   call,
@@ -11,6 +14,7 @@ import {
   SelectEffect,
   takeEvery,
 } from "redux-saga/effects";
+import { IAppointment } from "src/types/appointment";
 import { IEmployee } from "src/types/employee";
 import { getServicesByShopId } from "src/utils/shopUtil";
 
@@ -57,6 +61,70 @@ function* readShopEmployees(): Generator<
   }
 }
 
+function getAllAppointments(shopId: string): Promise<IAppointment[]> {
+  return fetch(`/api/shop/${shopId}/appointments/`, {
+    method: "GET",
+    headers: {
+      Accept: "application/json",
+      "Content-Type": "application/json",
+    },
+  }).then((res) => {
+    if (res.status === 200) {
+      return res.json().then((data) => {
+        const appointments = data
+          .map((appointment: Appointment) => {
+            const vehicleId = appointment.vehicle_id;
+            const customerId = appointment.customer_id;
+            const serviceId = appointment.service_id;
+
+            if (vehicleId && customerId && serviceId) {
+              const vehiclePromise = getVehicleById(vehicleId);
+              const customerPromise = getCustomerById(customerId);
+              const servicePromise = getServiceById(serviceId);
+
+              return Promise.all([
+                vehiclePromise,
+                customerPromise,
+                servicePromise,
+              ]).then((values) => {
+                const vehicle = values[0];
+                const customer = values[1];
+                const service = values[2];
+
+                return {
+                  id: appointment.id,
+                  startTime: appointment.start_time,
+                  endTime: appointment.end_time,
+                  customer: customer,
+                  shopId: appointment.shop_id,
+                  quoteId: appointment.quote_id,
+                  serviceName: service?.name,
+                  price: appointment.price,
+                  status: appointment.status,
+                  workOrderId: appointment.work_order_id,
+                  vehicle: vehicle,
+                };
+              });
+            } else {
+              // Return undefined if vehicle, customer, or service are null and filter out invalid appointments below
+              return;
+            }
+          })
+          .filter((appointment: IAppointment | undefined) => {
+            return appointment !== undefined;
+          });
+
+        return Promise.all(appointments).then((appointmentList) => {
+          return appointmentList;
+        });
+      });
+    } else {
+      // TODO: check and handle errors
+      return [];
+    }
+  });
+}
+
 function* readShopServices(): Generator<CallEffect | PutEffect | SelectEffect> {
   const shopId = (yield select(AuthSelectors.getShopId)) as string | null;
   if (shopId) {
@@ -68,6 +136,19 @@ function* readShopServices(): Generator<CallEffect | PutEffect | SelectEffect> {
   }
 }
 
+function* readShopAppointments(): Generator<
+  CallEffect | PutEffect | SelectEffect
+> {
+  const shopId = (yield select(AuthSelectors.getShopId)) as string | null;
+  if (shopId) {
+    const appointments = yield call(getAllAppointments, shopId);
+    yield put({
+      type: ShopTypes.SET_SHOP_APPOINTMENTS,
+      payload: { shopId, appointments },
+    });
+  }
+}
+
 /**
  * Saga to handle all employee related actions.
  */
@@ -75,5 +156,6 @@ export function* shopSaga() {
   yield all([
     takeEvery(ShopTypes.READ_SHOP_EMPLOYEES, readShopEmployees),
     takeEvery(ShopTypes.READ_SHOP_SERVICES, readShopServices),
+    takeEvery(ShopTypes.READ_SHOP_APPOINTMENTS, readShopAppointments),
   ]);
 }
