+++ src/redux/selectors/appointmentSelectors.ts
@@ -1,5 +1,5 @@
 import { createSelector } from "reselect";
-import { IAppointmentsState } from "../state/user/appointmentState";
+import { IAppointmentsState } from "../state/shop/appointmentState";
 import { RootState } from "../store";
 
 const getAppointmentsState = (state: RootState): IAppointmentsState =>
@@ -10,12 +10,6 @@ const getAppointments = createSelector(
   (appointmentsState) => appointmentsState.appointments
 );
 
-const getCustomerAppointments = createSelector(
-  getAppointmentsState,
-  (appointmentsState) => appointmentsState.customerAppointments
-);
-
 export const AppointmentSelectors = {
   getAppointments,
-  getCustomerAppointments,
 };
