+++ src/redux/types/appointmentTypes.ts
@@ -1,8 +1,7 @@
 enum AppointmentTypes {
   SET_APPOINTMENT_STATUS = "SET APPOINTMENT STATUS",
-  READ_APPOINTMENTS = "CREATE GET APPOINTMENTS",
-  READ_CUSTOMER_APPOINTMENTS = "CREATE GET CUSTOMER APPOINTMENTS",
-  SET_APPOINTMENTS = "SET APPOINTMENT LIST",
+  READ_APPOINTMENTS = "READ APPOINTMENTS",
+  SET_APPOINTMENTS = "SET APPOINTMENTS",
   CREATE_APPOINTMENT = "CREATE APPOINTMENT",
 }
