+++ src/redux/types/shopTypes.ts
@@ -4,6 +4,8 @@ enum ShopTypes {
   READ_SHOP_SERVICES = "READ SERVICES",
   SET_SHOP_SERVICES = "SET SERVICES",
   SET_SHOP_STATE = "SET SHOP STATE",
+  SET_SHOP_APPOINTMENTS = "SET APPOINTMENTS",
+  READ_SHOP_APPOINTMENTS = "READ APPOINTMENTS",
 }
 
 export default ShopTypes;
